Add M1 integration tests
The integeration tests (M1ErrorTest and M1RuntimeTest) test the example
programmes against the hole compiler chain.
The RuntimeTest need a LLVM installation to work, because they execute the
LLVM IR code and check for the exitCode.
In particular the lli tool is needed. The RuntimeTest checks for a working LLVM
installation, if no installation is found the tests are ignored.